+++ addons/taliban/CfgVehicles.hpp
@@ -9,11 +9,13 @@ class CfgVehicles {
     #include "Veterans\CfgVehicles_Veterans_I.hpp"
     #include "Veterans\CfgVehicles_Veterans_O.hpp"
     #include "Veterans\CfgVehicles_Turrets_Veterans.hpp"
-    #include "Veterans\CfgVehicles_Vehicles_Veterans.hpp"
+    #include "Veterans\CfgVehicles_Vehicles_Veterans_I.hpp"
+    #include "Veterans\CfgVehicles_Vehicles_Veterans_O.hpp"
 
     // Militia
     #include "Militia\CfgVehicles_Militia_I.hpp"
     #include "Militia\CfgVehicles_Militia_O.hpp"
     #include "Militia\CfgVehicles_Turrets_Militia.hpp"
-    #include "Militia\CfgVehicles_Vehicles_Militia.hpp"
+    #include "Militia\CfgVehicles_Vehicles_Militia_I.hpp"
+    #include "Militia\CfgVehicles_Vehicles_Militia_O.hpp"
 };
