+++ addons/taliban/Militia/CfgVehicles_Vehicles_Militia_O.hpp
@@ -1,0 +1,57 @@
+class TACU_Taliban_Militia_V_O_Kamaz_ZU23: TACU_Taliban_Militia_V_I_Kamaz_ZU23 {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+class TACU_Taliban_Militia_V_O_Kamaz_Transport: TACU_Taliban_Militia_V_I_Kamaz_Transport {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+class TACU_Taliban_Militia_V_O_Kamaz_Covered_Transport: TACU_Taliban_Militia_V_I_Kamaz_Covered_Transport {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+class TACU_Taliban_Militia_V_O_Truck: TACU_Taliban_Militia_V_I_Truck {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+class TACU_Taliban_Militia_V_O_Offroad_Unarmed: TACU_Taliban_Militia_V_I_Offroad_Unarmed {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+class TACU_Taliban_Militia_V_O_Offroad_HMG: TACU_Taliban_Militia_V_I_Offroad_HMG {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+class TACU_Taliban_Militia_V_O_Offroad_AT: TACU_Taliban_Militia_V_I_Offroad_AT {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+class TACU_Taliban_Militia_V_O_Pickup_Unarmed: TACU_Taliban_Veterans_V_O_Pickup_Unarmed {
+    faction = "TACU_Taliban_Militia_O";
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+class TACU_Taliban_Militia_V_O_Pickup_HMG: TACU_Taliban_Veterans_V_O_Pickup_HMG {
+    faction = "TACU_Taliban_Militia_O";
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+class TACU_Taliban_Militia_V_O_Pickup_MMG: TACU_Taliban_Veterans_V_O_Pickup_MMG {
+    faction = "TACU_Taliban_Militia_O";
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
