+++ addons/taliban/Veterans/CfgVehicles_Vehicles_Veterans_O.hpp
@@ -1,0 +1,132 @@
+class TACU_Taliban_Veterans_V_O_Kamaz_ZU23: TACU_Taliban_Veterans_V_I_Kamaz_ZU23 {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Kamaz_Transport: TACU_Taliban_Veterans_V_I_Kamaz_Transport {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Kamaz_Covered_Transport: TACU_Taliban_Veterans_V_I_Kamaz_Covered_Transport {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Truck: TACU_Taliban_Veterans_V_I_Truck {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Offroad_Unarmed: TACU_Taliban_Veterans_V_I_Offroad_Unarmed {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Offroad_HMG: TACU_Taliban_Veterans_V_I_Offroad_HMG {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Offroad_AT: TACU_Taliban_Veterans_V_I_Offroad_AT {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Offroad_Unarmed_UP: TACU_Taliban_Veterans_V_I_Offroad_Unarmed_UP {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Offroad_HMG_UP: TACU_Taliban_Veterans_V_I_Offroad_HMG_UP {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Offroad_AT_UP: TACU_Taliban_Veterans_V_I_Offroad_AT_UP {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Pickup_Unarmed: TACU_Taliban_Veterans_V_I_Pickup_Unarmed {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Pickup_unarmed_UP: TACU_Taliban_Veterans_V_I_Pickup_unarmed_UP {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Pickup_HMG: TACU_Taliban_Veterans_V_I_Pickup_HMG {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Pickup_HMG_UP: TACU_Taliban_Veterans_V_I_Pickup_HMG_UP {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Pickup_MMG: TACU_Taliban_Veterans_V_I_Pickup_MMG {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Pickup_MMG_UP: TACU_Taliban_Veterans_V_I_Pickup_MMG_UP {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_BRDM2: TACU_Taliban_Veterans_V_I_BRDM2 {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Hilux: TACU_Taliban_Veterans_V_I_Hilux {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Hilux_Armored: TACU_Taliban_Veterans_V_I_Hilux_Armored {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Hilux_BMP: TACU_Taliban_Veterans_V_I_Hilux_BMP {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_Hilux_BMP_Armored: TACU_Taliban_Veterans_V_I_Hilux_BMP {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+class TACU_Taliban_Veterans_V_O_T72: TACU_Taliban_Veterans_V_I_T72 {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
